# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PCMI API
  description: Persistent Cognitive Memory Infrastructure — memory API (v1)
  version: 1.51.0
servers:
  - url: http://localhost:8000
    description: Local
security:
  - ApiKeyAuth: []
paths:
...
  /v1/extraction-profiles/{profile_id}:
    put:
      summary: Create or update a tenant extraction profile
      parameters:
        - name: profile_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile]
              properties:
                path_prefix: { type: string, default: root }
                enabled: { type: boolean, default: true }
                profile:
                  type: object
                  description: Domain profile JSON (required_slots, entity_promotion, …)
      responses:
        "200":
          description: Stored profile
        "400":
          description: Invalid profile
    delete:
      summary: Delete an extraction profile
      parameters:
        - name: profile_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /v1/memories/extraction/{memory_id}:
    get:
      summary: Read pcmi_extract for a memory (current version)
      parameters:
        - name: memory_id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Extraction record or null
        "404":
          description: Memory not found
    post:
      summary: Force LLM attribute extraction for a memory
      parameters:
        - name: memory_id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Extraction completed
        "404":
          description: No memory or matching profile
        "422":
          description: LLM/validation failure (partial extraction may be returned)
        "503":
          description: EXTRACTION_ENABLED=false or LLM not configured
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...